requirements.txt
alabaster==1.0.0
alembic==1.13.2
auth==0.5.3
babel==2.16.0
bcrypt==4.2.0
blinker==1.8.2
boto3==1.35.2
botocore==1.35.2
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
DateTime==5.5
distlib==0.3.8
dnspython==2.6.1
docutils==0.21.2
email_validator==2.2.0
eventlet==0.36.1
falcon==3.1.3
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==23.0.0
hvac==2.3.0
identify==2.6.0
idna==3.7
imagesize==1.4.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
Mako==1.3.5
MarkupSafe==2.1.5
mongoengine==0.28.2
nodeenv==1.9.1
packaging==24.1
platformdirs==4.2.2
pre-commit==3.8.0
psycopg2-binary==2.9.9
Pygments==2.18.0
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.2
setuptools==73.0.1
six==1.16.0
snowballstemmer==2.2.0
Sphinx==8.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.32
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.26.3
Werkzeug==3.0.3
zope.interface==7.0.1
